+++ src/main/java/proteomics/Types/DebugEntry.java
@@ -4,10 +4,10 @@ public class DebugEntry implements Comparable<DebugEntry> {
 
     public final String chain;
     public final int link_site;
-    public final float mass;
+    public final double mass;
     public final double score;
 
-    public DebugEntry(String chain, int link_site, float mass, double score) {
+    public DebugEntry(String chain, int link_site, double mass, double score) {
         this.chain = chain;
         this.link_site = link_site;
         this.mass = mass;
